reverted morph.py back to dream.py
