+++ README.md
@@ -1,8 +1,3 @@
-# small to do list
-readline setup
-
-prompt morphing with weighting
-
 # Stable Diffusion Dream Script
 
 This is a fork of CompVis/stable-diffusion, the wonderful open source
